Proteeti/Proteeti/requirements.txt
Flask==3.0.3
gunicorn
flask-dance==7.1.0
Flask-Login
requests==2.31.0
python-dotenv==1.0.0
psycopg[binary]>=3.1.0
Flask-SQLAlchemy==3.1.1
bcrypt==4.1.2
flask-mail
Werkzeug
Flask-Migrate
flask-cors
psycopg2-binary